Update layout and home page structure for improved responsiveness and user experience
- Added viewport settings and favicon to the metadata in layout.tsx for better mobile support and branding.
- Refactored the home page to utilize a new HomeClient component, removing the previous image slideshow and header/footer structure for a cleaner layout.
- Introduced a MountedContext in providers.tsx to manage component mount state, enhancing the rendering logic for child components.
